feat(code): standardises input highlighting
