+++ Assets/Scripts/UI/Components/DiaryQABlock.cs
@@ -2,20 +2,7 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class DiaryQABlock : MonoBehaviour
+public class DiaryQABlock : InputHighlighter
 {
     public TextMeshProUGUI questionText;
-    public TMP_InputField inputField;
-    [SerializeField] private GameObject selectedOutline;
-
-    private void Update()
-    {
-        bool isSelected = UIManager.Instance.IsObjectSelected(inputField.gameObject);
-        bool isFocused = inputField.isFocused;
-
-        selectedOutline.SetActive(isSelected || isFocused);
-
-        if (isFocused && InputManager.Instance.GamepadIsCurrentInput()) selectedOutline.GetComponent<Image>().color = UIColors.white;
-        else if (isSelected) selectedOutline.GetComponent<Image>().color = UIColors.terminalGreen;
-    }
 }
