+++ Assets/Scripts/UI/Components/InputHighlighter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cd143d7ca65c57d49bef43d29bcabd49
